refactor: dto 이름 수정

+++ src/main/java/com/review/shop/dto/orders/OrderItemDTO.java
@@ -5,8 +5,8 @@ import lombok.Data;
 
 @Data
 public class OrderItemDTO {
-    private long order_item_id;
-    private long order_id;      // ORDERS insert 후 받아온 ID
+    private int order_item_id;
+    private int order_id;      // ORDERS insert 후 받아온 ID
     private int product_id;
     private String prd_name;
     private int product_price;
